# spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Puja Booking Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/me": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Me",
        "description": "Current admin session \u2014 roles from DB (not JWT) for accurate RBAC nav.",
        "operationId": "admin_me_v1_admin_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/admin/settings/advance-booking-amount": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Advance",
        "operationId": "get_advance_v1_admin_settings_advance_booking_amount_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceAmountResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Set Advance",
        "operationId": "set_advance_v1_admin_settings_advance_booking_amount_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceAmountUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceAmountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/admin/users/{user_id}/roles": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List User Roles",
        "operationId": "list_user_roles_v1_admin_users__user_id__roles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Assign Role",
        "operationId": "assign_role_v1_admin_users__user_id__roles_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/users/{user_id}/roles/{role}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Revoke Role",
        "operationId": "revoke_role_v1_admin_users__user_id__roles__role__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/users/{user_id}/credential": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Provision Credential",
        "description": "Generate (or reset) a target admin's TOTP secret. Secret shown ONCE.\n\nAdmin-vouched provisioning replaces phone-only self-enrolment \u2014 it closes\nthe enrolment race and keeps SMS/DLT out of the admin path entirely. Used\nfor onboarding a new admin and for resetting a lost authenticator.",
        "operationId": "provision_credential_v1_admin_users__user_id__credential_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialProvisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminMeResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "default": false
          },
          "is_support": {
            "type": "boolean",
            "title": "Is Support",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "phone",
          "roles"
        ],
        "title": "AdminMeResponse",
        "description": "Session identity for the admin shell (A-ADMIN-UI role-gated nav)."
      },
...
      "AdvanceAmountResponse": {
        "properties": {
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "currency"
        ],
        "title": "AdvanceAmountResponse"
      },
      "AdvanceAmountUpdate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "minimum": 1.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "change_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "AdvanceAmountUpdate"
      },
...
      "CredentialProvisionResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "provisioning_uri": {
            "type": "string",
            "title": "Provisioning Uri"
          },
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "Scan in an authenticator app now. This secret is shown only once."
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "provisioning_uri",
          "secret"
        ],
        "title": "CredentialProvisionResponse",
        "description": "Returned ONCE at provisioning \u2014 the URI/secret are never retrievable again."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleAssignRequest": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "support"
            ],
            "title": "Role"
          },
          "change_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "RoleAssignRequest"
      },
...
      "UserRolesResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "roles"
        ],
        "title": "UserRolesResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}